requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangGraph and LangChain


# Google AI and Web Search
google-generativeai==0.3.2
tavily-python==0.7.22

# Frontend and Database
streamlit==1.31.0
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Deployment
gunicorn==21.2.0
